assigment/requirements.txt
google-genai
google-cloud-storage
google-cloud-pubsub
google-cloud-redis
google-cloud-aiplatform
functions-framework
langgraph
